+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.6.0] - 2022-09-30
 ### Added
 - Warmup-1 in3050
+- Warmup-1 max1020
 
 ## [1.5.0] - 2022-09-29
 ### Added